refactor: consolidate test helper functions into distributed::test_helpers module
- Move create_test_drt_async and create_test_drt_with_settings_async from metrics and system_status_server modules to centralized location
- Update all test imports to use the consolidated helpers
- Eliminate code duplication across test modules
