+++ lib/runtime/src/metrics.rs
@@ -578,14 +578,7 @@ mod test_helpers {
     use super::prometheus_names::{nats_client, nats_service};
     use super::*;
 
-    /// Helper function to create a DRT instance for testing in async contexts
-    #[cfg(feature = "integration")]
-    pub async fn create_test_drt_async() -> crate::DistributedRuntime {
-        let rt = crate::Runtime::from_current().unwrap();
-        crate::DistributedRuntime::from_settings_without_discovery(rt)
-            .await
-            .unwrap()
-    }
+    // Helper function moved to crate::test_helpers::create_test_drt_async
 
     /// Base function to filter Prometheus output lines based on a predicate.
     /// Returns lines that match the predicate, converted to String.
@@ -927,7 +920,7 @@ mod test_metricsregistry_prefixes {
 
     #[tokio::test]
     async fn test_hierarchical_prefixes_and_parent_hierarchies() {
-        let drt = super::test_helpers::create_test_drt_async().await;
+        let drt = crate::distributed::test_helpers::create_test_drt_async().await;
 
         const DRT_NAME: &str = "";
         const NAMESPACE_NAME: &str = "ns901";
@@ -1002,7 +995,7 @@ mod test_metricsregistry_prefixes {
     #[tokio::test]
     async fn test_recursive_namespace() {
         // Create a distributed runtime for testing
-        let drt = super::test_helpers::create_test_drt_async().await;
+        let drt = crate::distributed::test_helpers::create_test_drt_async().await;
 
         // Create a deeply chained namespace: ns1.ns2.ns3
         let ns1 = drt.namespace("ns1").unwrap();
@@ -1060,7 +1053,7 @@ mod test_metricsregistry_prometheus_fmt_outputs {
     #[tokio::test]
     async fn test_prometheusfactory_using_metrics_registry_trait() {
         // Setup real DRT and registry using the test-friendly constructor
-        let drt = super::test_helpers::create_test_drt_async().await;
+        let drt = crate::distributed::test_helpers::create_test_drt_async().await;
 
         // Use a simple constant namespace name
         let namespace_name = "ns345";
@@ -1318,7 +1311,7 @@ mod test_metricsregistry_nats {
     #[tokio::test]
     async fn test_drt_nats_metrics() {
         // Setup real DRT and registry using the test-friendly constructor
-        let drt = super::test_helpers::create_test_drt_async().await;
+        let drt = crate::distributed::test_helpers::create_test_drt_async().await;
 
         // Get DRT output which should include NATS client metrics
         let drt_output = drt.prometheus_metrics_fmt().unwrap();
@@ -1382,7 +1375,7 @@ mod test_metricsregistry_nats {
         // the values of the metrics.
 
         // Setup real DRT and registry using the test-friendly constructor
-        let drt = super::test_helpers::create_test_drt_async().await;
+        let drt = crate::distributed::test_helpers::create_test_drt_async().await;
 
         // Create a namespace and components from the DRT
         let namespace = drt.namespace("ns789").unwrap();
